Add files via upload
This can be open through the pencil evolus app. Static Snapshots are provided to view as a reference. To ensure that navigation clicks properly, change the display resolution in your PC / Laptop to 100%. Excel files explain the errors we find through the hotel royal website and from there think of ways to further improve.